HRMS/requirements.txt
Django==6.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
